Refactoring: clean/merge security utility classes
