+++ leshan-server-core-demo/src/main/java/org/eclipse/leshan/server/core/demo/json/JacksonSecurityDeserializer.java
@@ -21,8 +21,8 @@ import java.security.GeneralSecurityException;
 import java.security.PublicKey;
 
 import org.eclipse.leshan.core.oscore.OscoreSetting;
+import org.eclipse.leshan.core.security.util.SecurityUtil;
 import org.eclipse.leshan.core.util.Hex;
-import org.eclipse.leshan.core.util.SecurityUtil;
 import org.eclipse.leshan.server.security.SecurityInfo;
 
 import com.fasterxml.jackson.core.JsonParseException;
